+++ src/main/java/store/orders/OrderRequest.java
@@ -7,7 +7,7 @@ import java.util.Date;
 public class OrderRequest {
     public Date orderDate;
     public String userName;
-    public Integer totalSum;
+    public Double totalSum;
     public String address;
     public ProductQuantity[] productQuantities;
 }
